requirements.txt
# ===================================================================================
#     Project: Image Classification Dataset
#     Description: Large image classification dataset with multiple categories 
#                  for training CNN models, transfer learning, and image recognition tasks.
#     
#     Author: Molla Samser
#     Email: help@rskworld.in
#     Phone: +91 93305 39277
#     Website: https://rskworld.in
#     
#     © 2025 RSK World. All rights reserved.
# ===================================================================================

# Core Dependencies
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.5.0

# Deep Learning Frameworks
tensorflow>=2.10.0
torch>=1.9.0
torchvision>=0.10.0

# Data Processing & Visualization
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0

# Image Augmentation
albumentations>=1.0.0
imgaug>=0.4.0

# Utilities
tqdm>=4.62.0
pyyaml>=6.0
python-dotenv>=0.19.0

# Jupyter Support
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=7.6.0

# Optional: GPU Support
# tensorflow-gpu>=2.10.0
# torch-cuda>=1.9.0

